requirements.txt
streamlit
google-genai
pillow
st-gsheets-connection
pandas